Updated bash script to work on other systems and add computer difficulty level
Prior to this the bash setup script would have to be changed for each user based on the current directory,
This script now gets the pwd and utilizes it along with relative directories to complete all commands.
Also updated some tests to the new status'

If the currentPlayer is a computer, sends the difficulty level along with it
This difficulty level is used to determine which clojure function to call.
One function gets a random move, the other gets the minimax move.